Changed pages directory (path prefix)
