+++ pages/admin/Rtl.tsx
@@ -19,6 +19,7 @@ import React from "react";
 // nodejs library that concatenates classes
 import classNames from "classnames";
 // react plugin used to create charts
+import { Line, Bar } from "react-chartjs-2";
 import {
   Chart as ChartJS,
   CategoryScale,
@@ -31,7 +32,6 @@ import {
   Filler,
   BarElement,
 } from "chart.js";
-import { Line, Bar } from "react-chartjs-2";
 
 ChartJS.register(
   CategoryScale,
@@ -71,9 +71,9 @@ import {
   chartExample2,
   chartExample3,
   chartExample4,
-} from "../variables/charts";
+} from "../../variables/charts";
 
-function Dashboard(props) {
+function Rtl() {
   const [bigChartData, setbigChartData] = React.useState("data1");
   const setBgChartData = (name) => {
     setbigChartData(name);
@@ -86,13 +86,13 @@ function Dashboard(props) {
             <Card className="card-chart">
               <CardHeader>
                 <Row>
-                  <Col className="text-left" sm="6">
-                    <h5 className="card-category">Total Shipments</h5>
-                    <CardTitle tag="h2">Performance</CardTitle>
+                  <Col className="text-right" sm="6">
+                    <h5 className="card-category">مجموع الشحنات</h5>
+                    <CardTitle tag="h2">أداء</CardTitle>
                   </Col>
                   <Col sm="6">
                     <ButtonGroup
-                      className="btn-group-toggle float-right"
+                      className="btn-group-toggle float-left"
                       data-toggle="buttons"
                     >
                       <Button
@@ -106,7 +106,7 @@ function Dashboard(props) {
                         onClick={() => setBgChartData("data1")}
                       >
                         <span className="d-none d-sm-block d-md-block d-lg-block d-xl-block">
-                          Accounts
+                          حسابات
                         </span>
                         <span className="d-block d-sm-none">
                           <i className="tim-icons icon-single-02" />
@@ -123,7 +123,7 @@ function Dashboard(props) {
                         onClick={() => setBgChartData("data2")}
                       >
                         <span className="d-none d-sm-block d-md-block d-lg-block d-xl-block">
-                          Purchases
+                          المشتريات
                         </span>
                         <span className="d-block d-sm-none">
                           <i className="tim-icons icon-gift-2" />
@@ -140,7 +140,7 @@ function Dashboard(props) {
                         onClick={() => setBgChartData("data3")}
                       >
                         <span className="d-none d-sm-block d-md-block d-lg-block d-xl-block">
-                          Sessions
+                          جلسات
                         </span>
                         <span className="d-block d-sm-none">
                           <i className="tim-icons icon-tap-02" />
@@ -162,18 +162,17 @@ function Dashboard(props) {
           </Col>
         </Row>
         <Row>
-          <Col lg="4">
+          <Col className="text-right" lg="4">
             <Card className="card-chart">
               <CardHeader>
-                <h5 className="card-category">Total Shipments</h5>
+                <h5 className="card-category">شحنات كاملة</h5>
                 <CardTitle tag="h3">
-                  <i className="tim-icons icon-bell-55 text-info" /> 763,215
+                  <i className="tim-icons icon-bell-55 text-primary" /> 763,215
                 </CardTitle>
               </CardHeader>
               <CardBody>
                 <div className="chart-area">
                   <Line
-                    datasetIdKey="id"
                     data={chartExample2.data()}
                     options={chartExample2.options}
                   />
@@ -181,12 +180,12 @@ function Dashboard(props) {
               </CardBody>
             </Card>
           </Col>
-          <Col lg="4">
+          <Col className="text-right" lg="4">
             <Card className="card-chart">
               <CardHeader>
-                <h5 className="card-category">Daily Sales</h5>
+                <h5 className="card-category">المبيعات اليومية</h5>
                 <CardTitle tag="h3">
-                  <i className="tim-icons icon-delivery-fast text-primary" />{" "}
+                  <i className="tim-icons icon-delivery-fast text-info" />{" "}
                   3,500€
                 </CardTitle>
               </CardHeader>
@@ -200,10 +199,10 @@ function Dashboard(props) {
               </CardBody>
             </Card>
           </Col>
-          <Col lg="4">
+          <Col className="text-right" lg="4">
             <Card className="card-chart">
               <CardHeader>
-                <h5 className="card-category">Completed Tasks</h5>
+                <h5 className="card-category">المهام المكتملة</h5>
                 <CardTitle tag="h3">
                   <i className="tim-icons icon-send text-success" /> 12,100K
                 </CardTitle>
@@ -220,18 +219,19 @@ function Dashboard(props) {
           </Col>
         </Row>
         <Row>
-          <Col lg="6" md="12">
-            <Card className="card-tasks">
-              <CardHeader>
-                <h6 className="title d-inline">Tasks(5)</h6>
-                <p className="card-category d-inline"> today</p>
-                <UncontrolledDropdown>
+          <Col className="text-center" lg="6" sm="6">
+            <Card className="card-tasks text-left">
+              <CardHeader className="text-right">
+                <h6 className="title d-inline">تتبع</h6>{" "}
+                <p className="card-category d-inline">اليوم</p>
+                <UncontrolledDropdown className="float-left">
                   <DropdownToggle
+                    aria-expanded={false}
+                    aria-haspopup={true}
                     caret
-                    className="btn-icon"
                     color="link"
                     data-toggle="dropdown"
-                    type="button"
+                    id="dropdownMenuLink"
                   >
                     <i className="tim-icons icon-settings-gear-63" />
                   </DropdownToggle>
@@ -240,19 +240,19 @@ function Dashboard(props) {
                       href="#pablo"
                       onClick={(e) => e.preventDefault()}
                     >
-                      Action
+                      عمل
                     </DropdownItem>
                     <DropdownItem
                       href="#pablo"
                       onClick={(e) => e.preventDefault()}
                     >
-                      Another action
+                      عمل آخر
                     </DropdownItem>
                     <DropdownItem
                       href="#pablo"
                       onClick={(e) => e.preventDefault()}
                     >
-                      Something else
+                      شيء آخر هنا
                     </DropdownItem>
                   </DropdownMenu>
                 </UncontrolledDropdown>
@@ -262,42 +262,7 @@ function Dashboard(props) {
                   <Table>
                     <tbody>
                       <tr>
-                        <td>
-                          <FormGroup check>
-                            <Label check>
-                              <Input defaultValue="" type="checkbox" />
-                              <span className="form-check-sign">
-                                <span className="check" />
-                              </span>
-                            </Label>
-                          </FormGroup>
-                        </td>
-                        <td>
-                          <p className="title">Update the Documentation</p>
-                          <p className="text-muted">
-                            Dwuamish Head, Seattle, WA 8:47 AM
-                          </p>
-                        </td>
-                        <td className="td-actions text-right">
-                          <Button
-                            color="link"
-                            id="tooltip636901683"
-                            title=""
-                            type="button"
-                          >
-                            <i className="tim-icons icon-pencil" />
-                          </Button>
-                          <UncontrolledTooltip
-                            delay={0}
-                            target="tooltip636901683"
-                            placement="right"
-                          >
-                            Edit Task
-                          </UncontrolledTooltip>
-                        </td>
-                      </tr>
-                      <tr>
-                        <td>
+                        <td className="text-center">
                           <FormGroup check>
                             <Label check>
                               <Input
@@ -311,35 +276,30 @@ function Dashboard(props) {
                             </Label>
                           </FormGroup>
                         </td>
-                        <td>
-                          <p className="title">GDPR Compliance</p>
-                          <p className="text-muted">
-                            The GDPR is a regulation that requires businesses to
-                            protect the personal data and privacy of Europe
-                            citizens for transactions that occur within EU
-                            member states.
-                          </p>
+                        <td className="text-right">
+                          <p className="title">مركز معالجة موقع محور</p>
+                          <p className="text-muted">نص آخر هناالوثائق</p>
                         </td>
-                        <td className="td-actions text-right">
+                        <td className="td-actions">
                           <Button
                             color="link"
-                            id="tooltip457194718"
+                            id="tooltip591536518"
                             title=""
                             type="button"
                           >
-                            <i className="tim-icons icon-pencil" />
+                            <i className="tim-icons icon-settings" />
                           </Button>
                           <UncontrolledTooltip
                             delay={0}
-                            target="tooltip457194718"
+                            target="tooltip591536518"
                             placement="right"
                           >
-                            Edit Task
+                            مهمة تحرير
                           </UncontrolledTooltip>
                         </td>
                       </tr>
                       <tr>
-                        <td>
+                        <td className="text-center">
                           <FormGroup check>
                             <Label check>
                               <Input defaultValue="" type="checkbox" />
@@ -349,33 +309,35 @@ function Dashboard(props) {
                             </Label>
                           </FormGroup>
                         </td>
-                        <td>
-                          <p className="title">Solve the issues</p>
+                        <td className="text-right">
+                          <p className="title">لامتثال GDPR</p>
                           <p className="text-muted">
-                            Fifty percent of all respondents said they would be
-                            more likely to shop at a company
+                            الناتج المحلي الإجمالي هو نظام يتطلب من الشركات
+                            حماية البيانات الشخصية والخصوصية لمواطني أوروبا
+                            بالنسبة للمعاملات التي تتم داخل الدول الأعضاء في
+                            الاتحاد الأوروبي.
                           </p>
                         </td>
-                        <td className="td-actions text-right">
+                        <td className="td-actions">
                           <Button
                             color="link"
-                            id="tooltip362404923"
+                            id="tooltip36890049"
                             title=""
                             type="button"
                           >
-                            <i className="tim-icons icon-pencil" />
+                            <i className="tim-icons icon-settings" />
                           </Button>
                           <UncontrolledTooltip
                             delay={0}
-                            target="tooltip362404923"
+                            target="tooltip36890049"
                             placement="right"
                           >
-                            Edit Task
+                            مهمة تحرير
                           </UncontrolledTooltip>
                         </td>
                       </tr>
                       <tr>
-                        <td>
+                        <td className="text-center">
                           <FormGroup check>
                             <Label check>
                               <Input defaultValue="" type="checkbox" />
@@ -385,32 +347,117 @@ function Dashboard(props) {
                             </Label>
                           </FormGroup>
                         </td>
-                        <td>
-                          <p className="title">Release v2.0.0</p>
+                        <td className="text-right">
+                          <p className="title">القضاياالقضايا</p>
                           <p className="text-muted">
-                            Ra Ave SW, Seattle, WA 98116, SUA 11:19 AM
+                            سيكونونقال 50٪ من جميع المستجيبين أنهم سيكونون أكثر
+                            عرضة للتسوق في شركة
                           </p>
                         </td>
-                        <td className="td-actions text-right">
+                        <td className="td-actions">
                           <Button
                             color="link"
-                            id="tooltip818217463"
+                            id="tooltip5456779"
                             title=""
                             type="button"
                           >
-                            <i className="tim-icons icon-pencil" />
+                            <i className="tim-icons icon-settings" />
                           </Button>
                           <UncontrolledTooltip
                             delay={0}
-                            target="tooltip818217463"
+                            target="tooltip5456779"
                             placement="right"
                           >
-                            Edit Task
+                            مهمة تحرير
                           </UncontrolledTooltip>
                         </td>
                       </tr>
                       <tr>
-                        <td>
+                        <td className="text-center">
+                          <FormGroup check>
+                            <Label check>
+                              <Input
+                                defaultChecked
+                                defaultValue=""
+                                type="checkbox"
+                              />
+                              <span className="form-check-sign">
+                                <span className="check" />
+                              </span>
+                            </Label>
+                          </FormGroup>
+                        </td>
+                        <td className="text-right">
+                          <p className="title">
+                            تصدير الملفات التي تمت معالجتها
+                          </p>
+                          <p className="text-muted">
+                            كما يبين التقرير أن المستهلكين لن يغفروا شركة بسهولة
+                            بمجرد حدوث خرق يعرض بياناتهم الشخصية.
+                          </p>
+                        </td>
+                        <td className="td-actions">
+                          <Button
+                            color="link"
+                            id="tooltip989428493"
+                            title=""
+                            type="button"
+                          >
+                            <i className="tim-icons icon-settings" />
+                          </Button>
+                          <UncontrolledTooltip
+                            delay={0}
+                            target="tooltip989428493"
+                            placement="right"
+                          >
+                            مهمة تحرير
+                          </UncontrolledTooltip>
+                        </td>
+                      </tr>
+                      <tr>
+                        <td className="text-center">
+                          <FormGroup check>
+                            <Label check>
+                              <Input
+                                defaultChecked
+                                defaultValue=""
+                                type="checkbox"
+                              />
+                              <span className="form-check-sign">
+                                <span className="check" />
+                              </span>
+                            </Label>
+                          </FormGroup>
+                        </td>
+                        <td className="text-right">
+                          <p className="title">الوصول إلى عملية التصدير</p>
+                          <p className="text-muted">
+                            سياسة السيء إنطلاق في قبل, مساعدة والمانيا أخذ قد.
+                            بل أما أمام ماشاء الشتاء،, تكاليف الإقتصادي بـ حين.
+                            ٣٠ يتعلّق للإتحاد ولم, وتم هناك مدينة بتحدّي إذ, كان
+                            بل عمل
+                          </p>
+                        </td>
+                        <td className="td-actions">
+                          <Button
+                            color="link"
+                            id="tooltip169784793"
+                            title=""
+                            type="button"
+                          >
+                            <i className="tim-icons icon-settings" />
+                          </Button>
+                          <UncontrolledTooltip
+                            delay={0}
+                            target="tooltip169784793"
+                            placement="right"
+                          >
+                            مهمة تحرير
+                          </UncontrolledTooltip>
+                        </td>
+                      </tr>
+                      <tr>
+                        <td className="text-center">
                           <FormGroup check>
                             <Label check>
                               <Input defaultValue="" type="checkbox" />
@@ -420,64 +467,29 @@ function Dashboard(props) {
                             </Label>
                           </FormGroup>
                         </td>
-                        <td>
-                          <p className="title">Export the processed files</p>
+                        <td className="text-right">
+                          <p className="title">الافراج عن v2.0.0</p>
                           <p className="text-muted">
-                            The report also shows that consumers will not easily
-                            forgive a company once a breach exposing their
-                            personal data occurs.
+                            عن رئيس طوكيو البولندي لمّ, مايو مرجع وباءت قبل هو,
+                            تسمّى الطريق الإقتصادي ذات أن. لغات الإطلاق الفرنسية
+                            دار ان, بين بتخصيص الساحة اقتصادية أم. و الآخ
                           </p>
                         </td>
-                        <td className="td-actions text-right">
+                        <td className="td-actions">
                           <Button
                             color="link"
-                            id="tooltip831835125"
+                            id="tooltip554497871"
                             title=""
                             type="button"
                           >
-                            <i className="tim-icons icon-pencil" />
+                            <i className="tim-icons icon-settings" />
                           </Button>
                           <UncontrolledTooltip
                             delay={0}
-                            target="tooltip831835125"
+                            target="tooltip554497871"
                             placement="right"
                           >
-                            Edit Task
-                          </UncontrolledTooltip>
-                        </td>
-                      </tr>
-                      <tr>
-                        <td>
-                          <FormGroup check>
-                            <Label check>
-                              <Input defaultValue="" type="checkbox" />
-                              <span className="form-check-sign">
-                                <span className="check" />
-                              </span>
-                            </Label>
-                          </FormGroup>
-                        </td>
-                        <td>
-                          <p className="title">Arival at export process</p>
-                          <p className="text-muted">
-                            Capitol Hill, Seattle, WA 12:34 AM
-                          </p>
-                        </td>
-                        <td className="td-actions text-right">
-                          <Button
-                            color="link"
-                            id="tooltip217595172"
-                            title=""
-                            type="button"
-                          >
-                            <i className="tim-icons icon-pencil" />
-                          </Button>
-                          <UncontrolledTooltip
-                            delay={0}
-                            target="tooltip217595172"
-                            placement="right"
-                          >
-                            Edit Task
+                            مهمة تحرير
                           </UncontrolledTooltip>
                         </td>
                       </tr>
@@ -487,62 +499,62 @@ function Dashboard(props) {
               </CardBody>
             </Card>
           </Col>
-          <Col lg="6" md="12">
+          <Col lg="6" sm="6">
             <Card>
-              <CardHeader>
-                <CardTitle tag="h4">Simple Table</CardTitle>
+              <CardHeader className="text-right">
+                <CardTitle tag="h4">جدول بسيط</CardTitle>
               </CardHeader>
               <CardBody>
                 <Table className="tablesorter" responsive>
                   <thead className="text-primary">
                     <tr>
-                      <th>Name</th>
-                      <th>Country</th>
-                      <th>City</th>
-                      <th className="text-center">Salary</th>
+                      <th>اسم</th>
+                      <th>بلد</th>
+                      <th>مدينة</th>
+                      <th className="text-center">راتب</th>
                     </tr>
                   </thead>
                   <tbody>
                     <tr>
-                      <td>Dakota Rice</td>
-                      <td>Niger</td>
-                      <td>Oud-Turnhout</td>
+                      <td>رايس داكوتا</td>
+                      <td>النيجر</td>
+                      <td>العود-تورنهاوت</td>
                       <td className="text-center">$36,738</td>
                     </tr>
                     <tr>
-                      <td>Minerva Hooper</td>
-                      <td>Curaçao</td>
-                      <td>Sinaai-Waas</td>
+                      <td>مينيرفا هوبر</td>
+                      <td>كوراساو</td>
+                      <td>Sinaai-واس</td>
                       <td className="text-center">$23,789</td>
                     </tr>
                     <tr>
-                      <td>Sage Rodriguez</td>
-                      <td>Netherlands</td>
-                      <td>Baileux</td>
+                      <td>سيج رودريجيز</td>
+                      <td>هولندا</td>
+                      <td>بايلي</td>
                       <td className="text-center">$56,142</td>
                     </tr>
                     <tr>
-                      <td>Philip Chaney</td>
-                      <td>Korea, South</td>
-                      <td>Overland Park</td>
+                      <td>فيليب شانيه</td>
+                      <td>كوريا، جنوب</td>
+                      <td>اوفرلاند بارك</td>
                       <td className="text-center">$38,735</td>
                     </tr>
                     <tr>
-                      <td>Doris Greene</td>
-                      <td>Malawi</td>
-                      <td>Feldkirchen in Kärnten</td>
+                      <td>دوريس غرين</td>
+                      <td>مالاوي</td>
+                      <td>المنع</td>
                       <td className="text-center">$63,542</td>
                     </tr>
                     <tr>
-                      <td>Mason Porter</td>
-                      <td>Chile</td>
-                      <td>Gloucester</td>
+                      <td>ميسون بورتر</td>
+                      <td>تشيلي</td>
+                      <td>غلوستر</td>
                       <td className="text-center">$78,615</td>
                     </tr>
                     <tr>
-                      <td>Jon Porter</td>
-                      <td>Portugal</td>
-                      <td>Gloucester</td>
+                      <td>جون بورتر</td>
+                      <td>البرتغال</td>
+                      <td>غلوستر</td>
                       <td className="text-center">$98,615</td>
                     </tr>
                   </tbody>
@@ -556,4 +568,4 @@ function Dashboard(props) {
   );
 }
 
-export default Dashboard;
+export default Rtl;
