+++ pages/index.tsx
@@ -3,7 +3,7 @@ import Router from "next/router";
 
 export default function Index() {
   React.useEffect(() => {
-    Router.push("/Dashboard");
+    Router.push("/admin/Dashboard");
   });
 
   return <div />;
